move "timetable" parsing to new file
